requirements.txt
bunch==1.0.1
py7zr==0.5.2
PyYAML==5.3
requests==2.18.4
urllib3==1.22
